working towards support for sidenotes (`[^label]:> content` vs. `[^label]: content` and `^[> inline content]` vs. `^[inline content]`) and section notes: the latter are footnotes which are shown at the (block boundary) location where they are defined in the text, rather than at the very end of the whole text.
This enables us to produce pages, which may be footnote-heavy a la the Gwern site, but with a choice of both sidenotes and 'inline, at the end of each BLOCK' "footnotes"/"pagenotes".

We choose to produce the footnotes at the spot where they are defined to ensure that the user has best control over their location in the final output and is not just dependent on any 'heuristics' like "print at end of section, or before start of next H1/H2/...", which would be much less flexible.